Move `determinePositionFromPath` to `XMLDiagnostics` as it is xml specific logic
